Issue 126
Enhancement: AGoodDirs, GoodDirsContinuousGen, LReachProblemDynamicsInterp, ExtIntEllApxBuilder, AMatrixOperations, CompositeMatrixOperations, MatrixODE45InterpFunc, MatrixSysNearestInterp, MatrixSysReshapeOde45RegInterp classes were changed. AMatrixBinaryOpArrayFunc, AMatrixUnaryOpArrayFunc, MatrixCatFunc, MatrixFlipdimFunc, MatrixReshapeFunc, MatrixSubArrayFunc, IMatrixSysFunction, MatrixODE45ScalarTimeInterpFunc classes were added.

I make new classes unary and binary operations for CompositeMatrixOperations. This classes designed to work with three-dimensional arrays. This is convenient, because it is not clear how to implement an operation cat() using the old classes. I also made ​​a new class MatrixODE45ScalarTimeInterpFunc of similar MatrixODE45InterpFunc. This class is designed to work with interpolating objects in the case of scalar time (length(timeVec) = 1). This is better than changing the function ode45 as in the latter case, we still can not get the interpolation object in the same form as for nonscalar time (this is due to the specific job VecOde45RegInterp).


git-svn-id: https://ellipsoids.googlecode.com/svn/branches/issue_126_vadimdanilov_2@2594 fbfa2205-8925-0410-8e01-63a555b7045b